refactor(notebooks): move plumbing constants and helpers to notebooks_helpers.go
Extract non-API-surface code out of notebooks_blob.go so the file focuses
on the CRUD methods and their public types. Moves the customHeadersKey /
NewRelic-Entity constants, the notebooksURL builder, blobRequestHeaders,
and the encodeEntityHeader marshaler (+ its swappable var) into a new
notebooks_helpers.go. Behaviour unchanged; unit tests and the integration
lifecycle test still pass with 100% coverage on hand-written code.
